+++ tests/Fixtures/InvalidPath/InvalidPathController.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Fixtures\InvalidPath;
+
+use Karewan\KnRoute\Attributes\Get;
+
+class InvalidPathController
+{
+	#[Get('missing-leading-slash')]
+	public function invalid(): void {}
+}